Reskin: account dropdown (#368)
* reskin account dropdown

* add backdrop, fix margins

* dynamic maxHeight value

* add account-padding-left var
