+++ extension/src/popup/views/Account/styles.scss
@@ -1,5 +1,9 @@
+:root {
+  --account-view-padding-left: 1.5rem;
+}
+
 .AccountView {
-  padding: 1.5rem;
+  padding: var(--account-view-padding-left);
 
   &__account-actions {
     align-items: center;
@@ -11,6 +15,8 @@
   &__name-key-display {
     flex-direction: row;
     width: 67%;
+    font-size: 0.875rem;
+    line-height: 1.375rem;
   }
 
   &__send-receive-display {
